+++ examples/state_mgmt/simple/lib/src/common.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class SomeWidget extends StatelessWidget {
+  final Widget child;
+
+  SomeWidget(this.child);
+
+  @override
+  Widget build(BuildContext context) {
+    return child;
+  }
+}
